+++ streamedrequests/responsesize.py
@@ -18,7 +18,7 @@
 from . import exceptions
 class SizeValidator:
     def __init__(self, max_size):
-        self.max_size = 0
+        self.max_size = max_size
         self.size = 0
 
     def add(self, amount):