+++ CHANGELOG.md
@@ -2,4 +2,14 @@
 
 This project uses Semantic Versioning
 
+## 1.0.0
+
+Post and get now return a tuple with a full requests response object and the response data
+
+## 0.0.1
+
+Fixed broken max_size validator, added tests for max_size validation
+
 ## 0.0.0
+
+Initial release